symbol_detection/requirements.txt
ultralytics
opencv-python
tqdm
pyyaml
